Support images
Previously we only supported markdown files (converting them to HTML).

Now we treat all files in the `content` directory as being intended for publication. The only exceptions are files and directories that begin with a dot ("."). These are ignored.

An exception is thrown if two handlers try to register for the same path.